Tests fix & separation
